+++ src/main/java/roomescape/controller/ReservationController.java
@@ -18,9 +18,9 @@ import roomescape.domain.Member;
 import roomescape.domain.MemberRole;
 import roomescape.service.MemberService;
 import roomescape.service.ReservationService;
-import roomescape.service.dto.ReservationMineResponse;
-import roomescape.service.dto.ReservationRequest;
-import roomescape.service.dto.ReservationResponse;
+import roomescape.service.dto.response.ReservationMineResponse;
+import roomescape.service.dto.request.ReservationRequest;
+import roomescape.service.dto.response.ReservationResponse;
 
 @RestController
 public class ReservationController {
