+++ src/main/java/roomescape/service/LoginService.java
@@ -8,10 +8,10 @@ import roomescape.domain.MemberRole;
 import roomescape.exception.login.InvalidTokenException;
 import roomescape.exception.login.UnauthorizedEmailException;
 import roomescape.exception.login.UnauthorizedPasswordException;
-import roomescape.service.dto.LoginCheckResponse;
-import roomescape.service.dto.LoginRequest;
-import roomescape.service.dto.SignupRequest;
-import roomescape.service.dto.SignupResponse;
+import roomescape.service.dto.response.LoginCheckResponse;
+import roomescape.service.dto.request.LoginRequest;
+import roomescape.service.dto.request.SignupRequest;
+import roomescape.service.dto.response.SignupResponse;
 import roomescape.service.helper.JwtTokenProvider;
 
 @Service
