+++ src/main/java/roomescape/service/dto/request/LoginRequest.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.request;
 
 public class LoginRequest {
     private final String email;
