+++ src/main/java/roomescape/service/dto/response/SignupResponse.java
@@ -1,4 +1,4 @@
-package roomescape.service.dto;
+package roomescape.service.dto.response;
 
 import roomescape.domain.Member;
 import roomescape.domain.MemberRole;
